♻️refactor: 훅 디렉토리 구조 통일 및 파일명 수정
